+++ hutool-extra/src/main/java/cn/hutool/extra/compress/Extractor.java
@@ -1,0 +1,4 @@
+package cn.hutool.extra.compress;
+
+public class Extractor {
+}